Refactor Skills module styles and remove max-width property
Refactor the Skills.module.scss file to remove the max-width property from the .skills_inner class. This change improves the responsiveness of the Skills component.

Move skillsData.js to assets/data folder

Update the import path for the skillsData module in the Skills.jsx file to reflect its new location in the assets/data folder. This change ensures that the correct data is imported for the Skills component.

Update PDF file paths in SideMenu component

Update the file paths for the PDF files in the SideMenu.jsx component to reflect the new location of the certificates folder. This change ensures that the correct PDF files are opened when the user clicks on the Resume and Certificates links.

Update dependencies in package.json

Update the @react-pdf-viewer/core and react-pdf dependencies in the package.json file to their latest versions. This change ensures that the project uses the most up-to-date versions of these dependencies.

Update ProjectCategories module styles

Update the ProjectCategories.module.scss file to use custom variables for background color and text color. This change improves the consistency and maintainability of the styles.

Add new certificate files

Add new certificate files to the public/certificates folder. These files include certificates for PMI Agile, Blue Prism Intelligent Automation, AWS EC2, AWS DevOps, and AWS CodeCommit.

Update routing in App.js

Update the routing in the App.js file to include a new Route for the Certificates component. This change ensures that the Certificates page is accessible to the user.

Update certificate file paths in Certificates component

Update the file paths for the certificate files in the Certificates.jsx component to reflect the new location of the certificates folder. This change ensures that the correct certificate files are displayed on the Certificates page.